:adhesive_bandage: Fix: price type int로 변경

+++ src/main/java/com/waterfogsw/gcccoffee/product/controller/dto/ProductAddRequest.java
@@ -5,14 +5,14 @@ import com.waterfogsw.gcccoffee.product.model.Category;
 public record ProductAddRequest(
         String name,
         Category category,
-        long price,
+        int price,
         String description
 ) {
     public ProductAddRequest {
         validate(name, category, price);
     }
 
-    private static void validate(String name, Category category, long price) {
+    private static void validate(String name, Category category, int price) {
         if (name == null || category == null || price <= 0) {
             throw new IllegalArgumentException();
         }
